examples/example_fast/demo-fast/requirements.txt
torch
transformers   
opencv-python
matplotlib
gradio
pillow
